Changed the way the next chunk is chosen
Refactored the code to change the way the next chunk is chosen for playback.
The mixed_data set is checked periodically for its emptiness. If true, check
again after 100ms, else return true
TODO : Currently, the output of the synchronizer cannot be played by VLC because
it reports that it is receiving incorrect headers. The headers are probably lost
while synchronizing. So we have to play initially, the content from one of the
peers. After synchronization, PlayChunk function can be called.